+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 ### Changed
   - Default key file name **(check when upgrading from an old version)**
   - ``keymsg-decrypt.py`` output file name
+  - Disable product output if no keys loaded
+  - Write single fill VCDU to packet file on VCID change
   - Rename FILL packets to IDLE packets
 
 ### Fixed